requirements.txt
ddt
pytest
openpyxl
selenium
unittest2
allure-pytest
webdriver-manager
allure-python-commons